Refactor, Resize all images coming to server to a fixed size before processing
The resizing part in this commit is not necessary as image sizes don't
affect LBPH recoginizer. We did it before we came to know that the image
sizes doesn't affect the result. For other info. on different
classifiers and different factors that affect result of those
classifiers refer OpenCV docs.